requirements.txt
Django==2.2
django-reversion==3.0.4
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.2
sqlparse==0.3.0
